requirements.txt
Flask==2.0.1
firebase-admin==5.3.0
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.1 